requirements.txt
langgraph==0.1.19
langchain==0.2.1
langchain-groq==0.1.4
langchain-core==0.2.1
chromadb==0.4.24
sentence-transformers==2.7.0
python-dotenv==1.0.1
pydantic==2.7.1